asdf

+++ test/test_ops.py
@@ -1151,7 +1151,7 @@ class TestOps(unittest.TestCase):
   def test_gather(self):
     # indices cannot have gradient
     b = torch.randint(3, size=[3,4,5,1,4], dtype=torch.int64, requires_grad=False)
-    a = Tensor(b.detach().numpy(), dtype=dtypes.int32, requires_grad=False)
+    a = Tensor(b.detach().numpy(), dtype=dtypes.float32, requires_grad=False)
     helper_test_op([(4,5,6,9,5)], lambda x: x.gather(index=b, dim=0), lambda x: x.gather(idx=a, dim=0))
     helper_test_op([(4,5,6,9,5)], lambda x: x.gather(index=b, dim=1), lambda x: x.gather(idx=a, dim=1))
     helper_test_op([(4,5,6,9,5)], lambda x: x.gather(index=b, dim=2), lambda x: x.gather(idx=a, dim=2))
